+++ packages/stream_video/CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased
+
+* Update stream_webrtc_flutter and device_info_plus dependencies
+
 ## 0.8.3
 
 ✅ Added