+++ packages/stream_video/CHANGELOG.md
@@ -3,6 +3,9 @@
 🐞 Fixed
 * (iOS) Fixed an issue where PiP will not stop when bringing app back to foreground
 
+✅ Added
+* Added `membersLimit` parameter to `getOrCreate()` and `join()` methods in `Call` class to limit number of members listed in the response.
+
 ## 0.8.3
 
 ✅ Added